Fix: front-js, dropdown-to-viewport fix
The new algorithm is based on the effective position of the expanding
element in the viewport, which is computed as the top offset of the
element minus the window scroll top.

Tested with my smartphone both when sticky-header on (scrolling or not)
and sticky-header off (scrolling or not)

p.s.
additionally a small typo fix, missing jquery alias ($) when trying
to animate html,body if stiky-header and no TCParams.dropdowntoViewport