Fix: allow smartload when jetpack's sharing enabled fixes #414
I used the solution proposed in the comment of the issue above.
Also removed a piece of code which I always found weird,
why re-build the img tag when no changes had to be made to it
instead of using matches[0] ?

Let me know if you have anything against that.