move after footer action to be after the row of widgets.
Move the action to happen after the div of the footer row of widgets is closed off - otherwise the "after_footer_widgets action happens in the same row.  Because the "colophon" is also a part of the footer, and we basically want to be before the colophon (but there isn't an action for that), moving this seems like the best option.