Fix: classic - fix slide textual elements filter prefix in admin filters like {prefix}_slide_(text|title|button|_length) are used both in front and in admin (when saving the slide img: CZR_meta_boxes class). In the classical style {prefix} is 'tc_' while in modern it's 'czr_'. The CZR_meta_boxes class is shared between classical and modern style.
So we should unify the filter prefixes in front too.
We could go with 'tc_' prefixes for both modern and classical relevant
filters, but this would create too much inconsistency in the modern
code, as I would prefer to not change all the modern slider (and non)
filter prefixes to avoid any snippet for the classical break the modern.

We could go with 'czr_' prefixes for the classical then, but this would
break the snippets compatibility.

Let's then apply the filters with the both prefixes in the meta boxes
class so that:
1) modern keeps working fine
2) snippets that worked for the classical will continue working.
fixes #1195