Fix: modern - fix RTL breadcrumb separator displayed with wrong direction fixes #1420
This seems to happen only in the modern style. In classical we don't
actually use a different separator in RTL: always Â» (see
http://www.glezer.co.il/2017/08/%D7%A9%D7%9C%D7%9E%D7%94-%D7%A9%D7%9C%D7%9E%D7%94-%D7%A9%D7%9C%D7%9E%D7%94-%D7%A9%D7%9C%D7%9E%D7%94-%D7%A9%D7%9C%D7%9E%D7%94/)
Basically what happens is that the browsers, recognizing that the word
that precedes the Â« or Â» is a "RTL set of characters", automatically
rotates the Â« or Â».  I propose then to use the same separator (Â»)
both in LTR and RTL, as we do in the classical style.